backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
boto3>=1.34.0
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
